// src/types.ts
export interface Session {
  user_id: string;
  user_name: string;
  role: 'rf' | 'supplier';
  supplier_id?: string;
}